Updates from AOCL-3.1

+++ gtests/sinh/sinh_callback.cc
@@ -54,12 +54,12 @@ double getFuncOp(double *data) {
   return LIBM_FUNC(sinh)(data[0]);
 }
 
-float getExpected(float *data) {
+double getExpected(float *data) {
   auto val = alm_mp_sinhf(data[0]);
   return val;
 }
 
-double getExpected(double *data) {
+long double getExpected(double *data) {
   auto val = alm_mp_sinh(data[0]);
   return val;
 }
